Add chat capability
This might be temporary.  If kept, it needs to evolve keep the history
longer and to share the web socket with the main communicator for game
states.  The real motivation is to start using a web socket based
server so we can sliminate polling.  Right now, there are two servers.
The original http-based one is used for game states and the web socket
based on is used for the chat.  Issue #5